FU-218 feat: adjust new product page
