+++ src/app/photographer/product/[productId]/page.tsx
@@ -1,7 +1,3 @@
-import ProductForm from "@/containers/photographer/product/product-form";
-
-const NewProduct = () => {
-  return <ProductForm />;
-};
-
-export default NewProduct;
+export default function ProductDetailPage() {
+  return <div />;
+}
