+++ src/containers/photographer/product/form/field/option-field-array.tsx
@@ -16,7 +16,7 @@ const OptionFieldArray = () => {
   });
 
   return (
-    <div style={{ width: "100%" }}>
+    <div style={{ minWidth: "fit-content" }}>
       <span className={formStyles.subtitle}>상품 옵션</span>
 
       {fields.map((item, index) => {
