+++ src/containers/photographer/product/form/index.tsx
@@ -15,45 +15,9 @@ import ImagesInput from "./field/image-input";
 import DiscountFieldArray from "./field/discount-field-array";
 import { formStyles } from "./form.css";
 
-const ProductForm = () => {
+const ProductForm = ({ formBase }: { formBase: ProductFormdata }) => {
   const router = useRouter();
-  const defaultValues: ProductFormdata = {
-    title: "",
-    subtitle: "",
-    items: [
-      {
-        title: "기본 가격",
-        content: "",
-        description: "",
-      },
-      {
-        title: "촬영 시간",
-        content: "1시간",
-        description: "",
-      },
-      {
-        title: "보정본 수",
-        content: "10장",
-        description: "보정본 추가는 상품 옵션에서 선택해 주세요.",
-      },
-    ],
-    options: [
-      {
-        title: "보정본 추가",
-        isFree: false,
-        description: "",
-        price: 0,
-      },
-    ],
-    discounts: [
-      {
-        title: "첫 주문 할인",
-        discountType: "AMOUNT",
-        discountValue: null,
-        description: "",
-      },
-    ],
-  };
+
   const productFormSchema = z.object({
     title: z
       .string()
@@ -142,7 +106,7 @@ const ProductForm = () => {
 
   const method = useForm<ProductFormdata>({
     resolver: zodResolver(productFormSchema),
-    defaultValues,
+    defaultValues: formBase,
   });
   const {
     handleSubmit,
